#7 Change project structure - Move components inside page folder - Since components are not intended to import pages in pages folder
