+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { HashRouter } from 'react-router-dom';
-import Footer from './components/Footer';
-import HeaderWithBodyRoute from './components/HeaderWithBodyRoute';
+import Footer from './pages/components/Footer';
+import HeaderWithBodyRoute from './pages/components/HeaderWithBodyRoute';
 
 function App() {
   return (
